Check if build/ exist before remove since it will fail if build/ doesn't exist

+++ code/raspberry-pi-5/build-arm64.sh
@@ -10,7 +10,9 @@ cp -r "$SCRIPT_DIR/../shared" "$SCRIPT_DIR/src/shared_real"
 # Build the cross-compile Docker image
 docker buildx build -f "$SCRIPT_DIR/Dockerfile.cross" --tag cross-pi .
 
-rm -r "$SCRIPT_DIR/build"
+if [ -d "$SCRIPT_DIR/build" ]; then
+    rm -r "$SCRIPT_DIR/build"
+fi
 
 # Build the project using the compile Dockerfile and output binaries to ./bin
 docker buildx build -f "$SCRIPT_DIR/Dockerfile.compile" -o type=local,dest="$SCRIPT_DIR/build" .
@@ -19,23 +21,23 @@ docker buildx build -f "$SCRIPT_DIR/Dockerfile.compile" -o type=local,dest="$SCR
 rm -rf "$SCRIPT_DIR/src/shared_real"
 
 jq --arg dir "$SCRIPT_DIR" \
-  '
+    '
   map(
     .directory |= gsub("^/code"; $dir) |
     .file |= gsub("^/code"; $dir) |
     .command |= gsub("/code"; $dir)
   )
-  ' \
-   "$SCRIPT_DIR"/build/compile_commands.json > "$SCRIPT_DIR"/build/compile_commands.json.tmp \
-   && mv "$SCRIPT_DIR"/build/compile_commands.json.tmp "$SCRIPT_DIR"/build/compile_commands.json
+' \
+    "$SCRIPT_DIR"/build/compile_commands.json > "$SCRIPT_DIR"/build/compile_commands.json.tmp \
+    && mv "$SCRIPT_DIR"/build/compile_commands.json.tmp "$SCRIPT_DIR"/build/compile_commands.json
 
 jq --arg sysroot "$SCRIPT_DIR/build/aarch64-linux-gnu/include" '
   map(
     .command |= gsub("/usr/aarch64-linux-gnu/include/"; ($sysroot + "/"))
   )
 ' "$SCRIPT_DIR/build/compile_commands.json" \
-  > "$SCRIPT_DIR/build/compile_commands.json.tmp" \
-  && mv "$SCRIPT_DIR/build/compile_commands.json.tmp" "$SCRIPT_DIR/build/compile_commands.json"
+    > "$SCRIPT_DIR/build/compile_commands.json.tmp" \
+    && mv "$SCRIPT_DIR/build/compile_commands.json.tmp" "$SCRIPT_DIR/build/compile_commands.json"
 
 # Generate .clangd with full absolute paths to cross headers
 SYSROOT="$SCRIPT_DIR/build/aarch64-linux-gnu/include"
